로그인, 회원가입, 로그아웃 v2

+++ dzhere-user-front/src/pages/client/check/check_index.js
@@ -1,31 +1,21 @@
-// Custom Navigation Drawer / Sidebar with Image and Icon in Menu Options
-// https://aboutreact.com/custom-navigation-drawer-sidebar-with-image-and-icon-in-menu-options/
-
 import * as React from 'react';
 import { useEffect } from 'react';
 import { View, Text, SafeAreaView, Alert, BackHandler } from 'react-native';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { logout } from '../../../modules/client/auth/auth';
-import { useSelector } from 'react-redux';
-import { useDispatch } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import client from '../../../lib/api/client';
+import { apiLogout } from '../../../lib/api/auth';
 
 const check_index = ({ navigation, route }) => {
-  // console.log(navigation);
-  // console.log(route);
-  // if (Platform.OS === "web") console.log(localStorage.getItem("user")); 
-  // else if (Platform.OS === "android")
-  //   console.log(asyncStorage.getItem("user")); 
   const dispatch = useDispatch();
   const {userInfo, } = useSelector(({auth}) => ({
     userInfo: auth.userInfo,
   }))
 
-
-  console.log('check_index 페이지에서 정보 받은 결과 : ', userInfo.token, typeof(userInfo.token));
-
-  // console.log('check_index 페이지에서 유저 토큰 받은 결과 : ', userToken);
+  console.log('check_index 페이지에서 받은 토큰 결과 : \n', userInfo.token, typeof(userInfo.token));
+  console.log('check_index 페이지에서 받은 헤더 정보 : \n', client.defaults.headers.common['Authorization']);
 
   // 뒤로가기 앱 종료
   const backAction = () => {
@@ -39,12 +29,12 @@ const check_index = ({ navigation, route }) => {
     ]);
     return true;
   };
-
   useEffect(() => {
     BackHandler.addEventListener('hardwareBackPress', backAction);
 
     return () => BackHandler.removeEventListener('hardwareBackPress', backAction);
   }, []);
+
   return (
     <SafeAreaView style={{ flex: 1, backgroundColor: "white" }}>
       <View style={{ flex: 1, padding: 16 }}>
@@ -65,58 +55,47 @@ const check_index = ({ navigation, route }) => {
             오늘의 출석
           </Text>
 
-          <Text>{String(userInfo.token)}</Text>
+          <Text>토큰 정보 : {'\n'+String(userInfo.token)}</Text>
           
-          {/* <Text>{}</Text> */}
           <TouchableOpacity
-            onPress={async () => {
-              try {
-                AsyncStorage.clear();
-                // if(Platform.OS == 'web'){
-                //   await localStorage.setItem('userToken', '');
-                //   await localStorage.setItem('userInfo', '');
-                // }
-                // if(Platform.OS=='android'){
-                //   AsyncStorage.clear();
-                //   // await AsyncStorage.setItem('userToken', '');
-                //   // await AsyncStorage.setItem('userInfo', '');
-                // }
-                dispatch(logout());
-                client.defaults.headers.common['Authorization'] = null;
-                console.log('로그아웃. Storage 초기화. client 헤더 초기화');
-                
-                navigation.navigate("UserLoginPage");
-                // navigation.reset({
-                //   index: 0,
-                //   routes: [
-                //     {
-                //       name: "UserLoginPage",
-                //     },
-                //   ],
-                // });
-              } catch (error) {
-                console.log('로그아웃 에러 : ', error);
-              }
-              
+            onPress={() => {
+                apiLogout()
+                .then(async (res) => {
+                  if(res.result){
+                    console.log('result : ',res.message);
+                    dispatch(logout());
+                    try{
+                      await AsyncStorage.clear();
+                      navigation.navigate("UserLoginPage");
+                    }
+                    catch (e) {
+                      console.log("Storage is not working : ", e);
+                    }
+                  } else{
+                    console.log(res.message);
+                  }
+                })
+                .catch((e) => {
+                  console.log("apiLogout.catch - e:", e);
+                });
             }}
           >
-            <Text>로그아웃</Text>
+            <Text style={{
+              fontSize: 25,
+              textAlign: "center",
+              marginBottom: 16,
+            }}>로그아웃</Text>
           </TouchableOpacity>
 
           <TouchableOpacity
-            // style={styles.signupBtn}
             onPress={() => client.get('api/user/test')}
           >
-            <Text>테스트</Text>
+            <Text style={{
+              fontSize: 25,
+              textAlign: "center",
+              marginBottom: 16,
+            }}>테스트</Text>
           </TouchableOpacity>
-          {/* <Button
-            onPress={() => navigation.navigate('SecondPage')}
-            title="Go to Second Page"
-          />
-          <Button
-            onPress={() => navigation.navigate('ThirdPage')}
-            title="Go to Third Page"
-          /> */}
         </View>
         <Text style={{ fontSize: 18, textAlign: "center", color: "grey" }}>
           Custom React Navigate Drawer
